+++ public/constants.js
@@ -1,14 +1,25 @@
 module.exports = {
     channels: {
+        //App
         MINIMIZE_APP: 'minimizeApp',
         MAXIMIZE_APP: 'maximizeApp',
         CLOSE_APP: 'closeApp',
+        //Tabs
         RELOAD_TAB: 'reloadTab',
         BACK_IN_TAB: 'backInTab',
         PREVIOUS_IN_TAB: 'previousInTab',
+        CLOSE_TAB: 'closeTab',
+        //Webview
         SEARCH_IN_TAB: 'searchInTab',
         OPEN_DEVTOOLS: 'openDevTools',
-        CLOSE_TAB: 'closeTab',
-        OPEN_URL_IN_NEW_TAB: 'openUrlInNewTab'
+        OPEN_URL_IN_NEW_TAB: 'openUrlInNewTab',
+        OPEN_URL: 'openURL',
+        OPEN_URL_IN_RENDERER: 'openURLInRenderer',
+        //UserPreferences
+        GET_USER_PREFERENCES: 'getUserPreferences',
+        //Bookmarks
+        GET_BOOKMARKS: 'getBookmarks',
+        ADD_BOOKMARK: 'addBookmark',
+        REMOVE_BOOKMARK: 'removeBookmark',
     },
 };