+++ src/components/Bookmark/BookmarkManager.jsx
@@ -1,7 +1,27 @@
+/* eslint-disable array-callback-return */
 import React from 'react';
+import { useSelector } from 'react-redux';
+import BookmarkFolder from './BookmarkFolder';
+import BookmarkItem from './BookmarkItem';
 
 export default function BookmarkManager() {
+    const bookmarks = useSelector(state => state.bookmarksReducer)
     return (
-        <div className='bookmarksManager'>Cette application utilise Chrome (v{window.versions.chrome()}), Node.js ({window.versions.node()}), et Electron ({window.versions.electron()})</div>
+        <div className='bookmarksManager'>
+            {bookmarks.map((bookmark, index) => {
+                if (bookmark.type === 'folder') {
+                    return <BookmarkFolder b={bookmark} key={index} />
+                } else if (bookmark.type === 'website') {
+                    return <BookmarkItem b={bookmark} key={index} />
+                }
+            })}
+
+            {
+                bookmarks.length === 0 &&
+                <div className='empty'>
+                    <p>There are no bookmarks yet.</p>
+                </div>
+            }
+        </div>
     )
 }