+++ src/components/Modals/AlertModal.jsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export default function AlertModal() {
+    return (
+        <div>AlertModal</div>
+    )
+}