+++ src/components/Tab/TabSystem.jsx
@@ -1,61 +1,75 @@
 import React, { useEffect, useState } from 'react'
 import { useSelector } from 'react-redux'
+import Searchbar from '../SearchBar/Searchbar'
+import { isEmpty } from '../utils'
 
 export default function TabSystem({ tabId }) {
     const [canGoBack, setCanGoBack] = useState(false)
     const [canGoForward, setCanGoForward] = useState(false)
-    const [isNewTab, setIsNewTab] = useState(true)
 
-    const [searchValue, setSearchValue] = useState("")
+    const [favicon, setFavicon] = useState(null)
+    const [isFavorited, setIsFavorited] = useState(false)
+
+    const [pageLoaded, setPageLoaded] = useState(false)
 
     const webviews = useSelector(state => state.webviewReducer)
+    const bookmarks = useSelector(state => state.bookmarksReducer)
     const [ref, setWebViewRef] = useState(null)
 
 
     useEffect(() => {
         const webViewRef = webviews.find(webview => webview.tabId === tabId)
 
-        const input = document.getElementById("searchInput");
-        if (isNewTab) {
-            input.focus();
-            input.select();
-            setIsNewTab(false)
-        }
-        input.addEventListener('focusin', () => {
-            input.select();
-        })
-
         if (webViewRef && webViewRef.webView.current) {
-            const event = webViewRef.webView.current;
             setWebViewRef(webViewRef.webView.current)
             setCanGoBack(webViewRef.webView.current.canGoBack())
             setCanGoForward(webViewRef.webView.current.canGoForward())
 
-            setSearchValue(event.getURL())
+            webViewRef.webView.current.addEventListener('page-favicon-updated', (e) => {
+                setFavicon(e.favicons[0])
+            });
 
-            event.addEventListener('did-navigate-in-page', (e) => {
-                setSearchValue(e.url)
-            })
+            const bookmark = bookmarks.find(bookmark => bookmark.url === webViewRef.webView.current.getURL())
+            if (!isEmpty(bookmark)) {
+                setIsFavorited(true)
+            } else {
+                setIsFavorited(false)
+            }
 
-            event.addEventListener('did-navigate', (e) => {
-                setSearchValue(e.url)
-            })
+            webViewRef.webView.current.addEventListener('did-start-loading', (e) => {
+                setPageLoaded(false)
+                const bookmark = bookmarks.find(bookmark => bookmark.url === webViewRef.webView.current.getURL())
+                if (!isEmpty(bookmark)) {
+                    setIsFavorited(true)
+                } else {
+                    setIsFavorited(false)
+                }
+
+            });
+
+            webViewRef.webView.current.addEventListener('did-stop-loading', (e) => {
+                setPageLoaded(true)
+            });
         }
 
         return () => {
-            console.log("unmount");
             if (webViewRef && webViewRef.webView.current) {
-                const event = webViewRef.webView.current;
-                event.removeEventListener('did-navigate-in-page', (e) => {
-                    setSearchValue(e.url)
-                })
-                event.removeEventListener('did-navigate', (e) => {
-                    setSearchValue(e.url)
-                })
+                webViewRef.webView.current.removeEventListener('page-favicon-updated', (e) => {
+                    setFavicon(e.favicons[0])
+                });
+
+                webViewRef.webView.current.removeEventListener('did-start-loading', (e) => {
+                    setPageLoaded(false)
+                });
+
+                webViewRef.webView.current.removeEventListener('did-stop-loading', (e) => {
+                    setPageLoaded(true)
+                });
+
             }
-            // const input = document.getElementById("searchInput");
         }
-    }, [webviews, isNewTab, tabId])
+    }, [webviews, tabId, bookmarks])
+
 
     const handleGoBack = () => {
         if (ref) {
@@ -75,29 +89,18 @@ export default function TabSystem({ tabId }) {
         }
     }
 
-    const parseURL = (url) => {
-        if (/^[^/]+\.[^/]+$/.test(url)) {
-            url = "https://" + url;
-        }
-
-        // Si l'entrée ressemble à une URL valide, naviguer vers cette URL.
-        if (/^(ftp|http|https):\/\/[^ "]+$/.test(url)) {
-            return url;
-        } else {
-            // Sinon, faire une recherche avec le moteur de recherche par défaut.
-            const searchQuery = encodeURIComponent(url);
-            const searchUrl = `https://www.google.com/search?q=${searchQuery}`;
-            return searchUrl;
-        }
+    const handleAddBookmark = () => {
+        window.bookmarks.addBookmark({
+            type: 'website',
+            name: ref.getTitle(),
+            favicon: favicon ? favicon : './img/icons/website.svg',
+            folder: null,
+            url: ref.getURL()
+        })
     };
 
-    const handleSearch = (e) => {
-        if (e.key === 'Enter') {
-            const url = parseURL(e.target.value)
-            if (ref) {
-                ref.loadURL(url)
-            }
-        }
+    const handlerRemoveBookmark = () => {
+        window.bookmarks.removeBookmark(ref.getURL())
     }
 
     return (
@@ -115,10 +118,15 @@ export default function TabSystem({ tabId }) {
                     </button>
                 </div>
                 <div className="searchBar">
-                    <button className='favoriteBTN'>
-                        <img src='./img/icons/favorite.svg' width={20} alt="favorite" />
+                    <button className='favoriteBTN' onClick={() => isFavorited ? handlerRemoveBookmark() : handleAddBookmark()} disabled={!pageLoaded}>
+                        {
+                            isFavorited ?
+                                <img src='./img/icons/favorite-filled.svg' width={20} alt="favorite" />
+                                :
+                                <img src='./img/icons/favorite.svg' width={20} alt="favorite" />
+                        }
                     </button>
-                    <input type="text" onKeyDown={handleSearch} onChange={(e) => setSearchValue(e.target.value)} value={searchValue} id="searchInput" />
+                    <Searchbar tabId={tabId} />
                 </div>
                 <div className="utilsButtons">
                     <button className='utilsBTN'>
@@ -130,7 +138,6 @@ export default function TabSystem({ tabId }) {
                     <button className='utilsBTN'>
                         <img src='./img/icons/expand.svg' width={20} alt="expand" />
                     </button>
-
                 </div>
             </div>
         </div>