+++ changelog.md
@@ -13,3 +13,19 @@
         - App icon
         - Change the IPC event to channels in constants file
         - Remove window.require and pass all IPC to preload.
+
+## 0.1.2:
+    - Fix:
+        - Sound Playing icon in tabBar display when no sound is playing
+        - Remove electron update for now
+    - Add:
+        - Bookmarks
+        - Click on BookmarkItem to open the bookmark
+        - Third Mouse Button to open bookmark in new tab
+        - userPrefernces in appData
+        - bookmarks in appData
+        - Store class to store userPreferences and bookmarks and later history
+        - Store last size of the window
+        - SearchBar
+        - Setup Sentry for Issues & Performance
+        - Setup DSN in env file